Refactor LoginForm
