+++ frontend/src/components/LoginForm.js
@@ -6,7 +6,6 @@
  * @requires src/services/loginService
  */
 import React from 'react'
-import loginService from '../services/loginService'
 import { InputText } from 'primereact/inputtext'
 import { Password } from 'primereact/password'
 import { Button } from 'primereact/button'
@@ -24,47 +23,11 @@ import { Button } from 'primereact/button'
  * @param {Function} param0.setUsername - Function to set the value of username
  * @param {string} param0.password - Value of password in form
  * @param {Function} param0.setPassword - Function to set the value of password
- * @param {object} param0.user - Object that contains details about current logged user
- * @param {Function} param0.setUser - Function that sets the value of user
+ * @param {object} param0.user - User details
+ * @param {Function} param0.handleLogin - Event handler for login
  * @returns {object} - Form for login and button for logout in JSX
  */
-const LoginForm = ({ username, setUsername, password, setPassword, user, setUser }) => {
-
-  /**
-   * Handle login button presses
-   *
-   * @type {object}
-   * @function
-   * @constant
-   * @memberof module:src/components/LoginForm
-   * @inner
-   * @param {object} event - Contains event
-   */
-  const handleLogin = async (event) => {
-    event.preventDefault()
-
-    try {
-      const user = await loginService.login({ username, password })
-      if (user.user) {
-        const appUser = {
-          username: user.user.username,
-          idToken: user.user.signInUserSession.idToken.jwtToken,
-          organisation: user.user.attributes['custom:organisation'],
-          admin: user.user.attributes['custom:admin']
-        }
-        setUser(appUser)
-        window.localStorage.setItem(
-          'loggedUser', JSON.stringify(appUser)
-        )
-      } else {
-        setUser(undefined)
-      }
-      setUsername('')
-      setPassword('')
-    } catch (exception) {
-      return
-    }
-  }
+const LoginForm = ({ username, setUsername, password, setPassword, user, handleLogin }) => {
 
   const containerStyle = {
     position: 'relative',
